Refactor : refactor pacakge structure (2)
