+++ src/test/java/com/palmseung/modules/admin/service/AdminServiceTest.java
@@ -1,14 +1,13 @@
 package com.palmseung.modules.admin.service;
 
 import com.palmseung.modules.members.domain.Member;
-import com.palmseung.modules.members.domain.MemberRepository;
+import com.palmseung.modules.members.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 import static com.palmseung.modules.members.MemberConstant.TEST_MEMBER;
 import static org.mockito.ArgumentMatchers.any;
@@ -21,18 +20,15 @@ public class AdminServiceTest {
     private AdminService adminService;
 
     @Mock
-    private MemberRepository memberRepository;
+    private MemberService memberService;
 
-    @Mock
-    private PasswordEncoder passwordEncoder;
-
-    @DisplayName("어드민 계정 생성 요청 시, 회원 저장 레포지토리 메소드가 호출된다.")
+    @DisplayName("어드민 계정 생성 요청 시, memberService의 create 메소드가 호출된다.")
     @Test
     void createAdminAccount() {
         //when
         adminService.createAdmin(TEST_MEMBER);
 
         //then
-        verify(memberRepository, times(1)).save(any(Member.class));
+        verify(memberService, times(1)).create(any(Member.class));
     }
 }
