+++ src/test/java/com/palmseung/modules/youtube/YouTubeVideosTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static com.palmseung.modules.youtube.support.YoutubeConstant.*;
+import static com.palmseung.modules.youtube.YoutubeConstant.*;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
